Allow expires_at to be a UNIX timestamp
This library currently only handles expires_at properties well when they
are provided in ISOString representation.

Unlike the expires_in field, the expires_at is not part of the OAuth2
spec, so implementations may vary.

Some APIs like the Strava API use a UNIX timestamp (seconds since epoch)
in the expires_at field:
http://developers.strava.com/docs/authentication/#response-parameters-1

Token refresh will yield "Invalid Date" in the expires_at field when
only accepting an ISOString.

This patch checks if expires_at is a number and if so assumes it is a
UNIX timestamp (seconds since epoch).